LICENSE: GPL v3 from https://www.gnu.org/licenses/gpl-3.0.txt
All the files say, "You should have received a copy of the GNU General
Public License along with this program." So let's give them a copy!

Also, GitHub should to detect this and add the license to its metadata for
the project so it's displayed on the project page and found in searches
such as `8088 license:gpl-3.0`. (Right now that finds just forks of this
repo that have added LICENSE.txt.)